requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Django==4.0.1
djangorestframework==3.13.1
dlib==19.22.99
et-xmlfile==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==3.3
imageio==2.14.0
jieba==0.42.1
networkx==2.6.3
numpy==1.22.1
opencv-python==4.5.5.62
openpyxl==3.0.9
packaging==21.3
pandas==1.4.0
Pillow==9.0.0
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
requests==2.27.1
scikit-image==0.19.1
scipy==1.7.3
six==1.16.0
sqlparse==0.4.2
tifffile==2021.11.2
tk==0.1.0
urllib3==1.26.8
xlrd==2.0.1
